refactor(azure): split provider into multiple files
This change restructures the Azure provider implementation by splitting it into multiple specialized files similar to the OpenAI implementation pattern.

- Reorganized into type definitions, defaults, utilities, and provider implementations

- Added proper exports in index.ts

- Updated imports and fixed circular dependencies

- Maintained backwards compatibility

- Fixed linting issues

- Added corresponding test structure
