+++ src/providers/azure/completion.ts
@@ -1,0 +1,115 @@
+import { fetchWithCache } from '../../cache';
+import { getEnvFloat, getEnvInt, getEnvString } from '../../envars';
+import logger from '../../logger';
+import type { CallApiContextParams, CallApiOptionsParams, ProviderResponse } from '../../types';
+import invariant from '../../util/invariant';
+import { REQUEST_TIMEOUT_MS } from '../shared';
+import { DEFAULT_AZURE_API_VERSION } from './defaults';
+import { AzureGenericProvider } from './generic';
+import { calculateAzureCost } from './util';
+
+export class AzureCompletionProvider extends AzureGenericProvider {
+  async callApi(
+    prompt: string,
+    context?: CallApiContextParams,
+    callApiOptions?: CallApiOptionsParams,
+  ): Promise<ProviderResponse> {
+    await this.ensureInitialized();
+    invariant(this.authHeaders, 'auth headers are not initialized');
+
+    if (!this.getApiBaseUrl()) {
+      throw new Error('Azure API host must be set.');
+    }
+
+    let stop: string;
+    try {
+      const stopEnvVar = getEnvString('OPENAI_STOP');
+      stop = stopEnvVar ? JSON.parse(stopEnvVar) : (this.config.stop ?? '');
+    } catch (err) {
+      throw new Error(`OPENAI_STOP is not a valid JSON string: ${err}`);
+    }
+    const body = {
+      model: this.deploymentName,
+      prompt,
+      max_tokens: this.config.max_tokens ?? getEnvInt('OPENAI_MAX_TOKENS', 1024),
+      temperature: this.config.temperature ?? getEnvFloat('OPENAI_TEMPERATURE', 0),
+      top_p: this.config.top_p ?? getEnvFloat('OPENAI_TOP_P', 1),
+      presence_penalty: this.config.presence_penalty ?? getEnvFloat('OPENAI_PRESENCE_PENALTY', 0),
+      frequency_penalty:
+        this.config.frequency_penalty ?? getEnvFloat('OPENAI_FREQUENCY_PENALTY', 0),
+      best_of: this.config.best_of ?? getEnvInt('OPENAI_BEST_OF', 1),
+      ...(stop ? { stop } : {}),
+      ...(this.config.passthrough || {}),
+    };
+
+    logger.debug(`Calling Azure API: ${JSON.stringify(body)}`);
+    let data;
+    let cached = false;
+    try {
+      ({ data, cached } = (await fetchWithCache(
+        `${this.getApiBaseUrl()}/openai/deployments/${
+          this.deploymentName
+        }/completions?api-version=${this.config.apiVersion || DEFAULT_AZURE_API_VERSION}`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            ...this.authHeaders,
+          },
+          body: JSON.stringify(body),
+        },
+        REQUEST_TIMEOUT_MS,
+      )) as any);
+    } catch (err) {
+      return {
+        error: `API call error: ${String(err)}`,
+      };
+    }
+
+    logger.debug(`Azure API response: ${JSON.stringify(data)}`);
+    try {
+      // Check for content filter
+      const choice = data.choices[0];
+      let output = choice.text;
+      if (output == null) {
+        if (choice.finish_reason === 'content_filter') {
+          output =
+            "The generated content was filtered due to triggering Azure OpenAI Service's content filtering system.";
+        } else {
+          output = '';
+        }
+      }
+
+      return {
+        output,
+        tokenUsage: cached
+          ? { cached: data.usage.total_tokens, total: data.usage.total_tokens }
+          : {
+              total: data.usage.total_tokens,
+              prompt: data.usage.prompt_tokens,
+              completion: data.usage.completion_tokens,
+            },
+        cost: calculateAzureCost(
+          this.deploymentName,
+          this.config,
+          data.usage?.prompt_tokens,
+          data.usage?.completion_tokens,
+        ),
+      };
+    } catch (err) {
+      return {
+        error: `API response error: ${String(err)}: ${JSON.stringify(data)}`,
+        tokenUsage: cached
+          ? {
+              cached: data.usage.total_tokens,
+              total: data.usage.total_tokens,
+            }
+          : {
+              total: data?.usage?.total_tokens,
+              prompt: data?.usage?.prompt_tokens,
+              completion: data?.usage?.completion_tokens,
+            },
+      };
+    }
+  }
+}
