+++ src/providers/azure/defaults.ts
@@ -1,0 +1,112 @@
+import type { AzureModelCost } from './types';
+
+export const DEFAULT_AZURE_API_VERSION = '2024-12-01-preview';
+
+export const AZURE_MODELS: AzureModelCost[] = [
+  {
+    id: 'gpt-4o-2024-08-06',
+    cost: {
+      input: 2.5 / 1000000,
+      output: 10 / 1000000,
+    },
+  },
+  {
+    id: 'gpt-4o',
+    cost: {
+      input: 5 / 1000000,
+      output: 15 / 1000000,
+    },
+  },
+  {
+    id: 'gpt-4o-mini',
+    cost: {
+      input: 0.15 / 1000000,
+      output: 0.6 / 1000000,
+    },
+  },
+  {
+    id: 'gpt-3.5-turbo-0125',
+    cost: {
+      input: 0.5 / 1000000,
+      output: 1.5 / 1000000,
+    },
+  },
+  {
+    id: 'gpt-3.5-turbo-instruct',
+    cost: {
+      input: 1.5 / 1000000,
+      output: 2 / 1000000,
+    },
+  },
+  {
+    id: 'gpt-4',
+    cost: {
+      input: 30 / 1000000,
+      output: 60 / 1000000,
+    },
+  },
+  {
+    id: 'gpt-4-32k',
+    cost: {
+      input: 60 / 1000000,
+      output: 120 / 1000000,
+    },
+  },
+  {
+    id: 'babbage-002',
+    cost: {
+      input: 0.4 / 1000000,
+      output: 0.4 / 1000000,
+    },
+  },
+  {
+    id: 'davinci-002',
+    cost: {
+      input: 2 / 1000000,
+      output: 2 / 1000000,
+    },
+  },
+  {
+    id: 'text-embedding-ada-002',
+    cost: {
+      input: 0.1 / 1000000,
+      output: 0.1 / 1000000,
+    },
+  },
+  {
+    id: 'text-embedding-3-large',
+    cost: {
+      input: 0.13 / 1000000,
+      output: 0.13 / 1000000,
+    },
+  },
+  {
+    id: 'text-embedding-3-small',
+    cost: {
+      input: 0.02 / 1000000,
+      output: 0.02 / 1000000,
+    },
+  },
+  // Legacy models
+  {
+    id: 'gpt-3.5-turbo-0301',
+    cost: {
+      input: 2 / 1000000,
+      output: 2 / 1000000,
+    },
+  },
+  {
+    id: 'gpt-3.5-turbo-0613',
+    cost: {
+      input: 1.5 / 1000000,
+      output: 2 / 1000000,
+    },
+  },
+  {
+    id: 'gpt-3.5-turbo-1106',
+    cost: {
+      input: 1 / 1000000,
+      output: 2 / 1000000,
+    },
+  },
+];
