+++ src/providers/azure/embedding.ts
@@ -1,0 +1,82 @@
+import { fetchWithCache } from '../../cache';
+import logger from '../../logger';
+import type { ProviderEmbeddingResponse } from '../../types';
+import invariant from '../../util/invariant';
+import { DEFAULT_AZURE_API_VERSION } from './defaults';
+import { AzureGenericProvider } from './generic';
+import { REQUEST_TIMEOUT_MS } from '../shared';
+
+export class AzureEmbeddingProvider extends AzureGenericProvider {
+  async callEmbeddingApi(text: string): Promise<ProviderEmbeddingResponse> {
+    await this.ensureInitialized();
+    invariant(this.authHeaders, 'auth headers are not initialized');
+    if (!this.getApiBaseUrl()) {
+      throw new Error('Azure API host must be set.');
+    }
+
+    const body = {
+      input: text,
+      model: this.deploymentName,
+    };
+    let data,
+      cached = false;
+    try {
+      ({ data, cached } = (await fetchWithCache(
+        `${this.getApiBaseUrl()}/openai/deployments/${this.deploymentName}/embeddings?api-version=${
+          this.config.apiVersion || DEFAULT_AZURE_API_VERSION
+        }`,
+        {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            ...this.authHeaders,
+          },
+          body: JSON.stringify(body),
+        },
+        REQUEST_TIMEOUT_MS,
+      )) as unknown as any);
+    } catch (err) {
+      return {
+        error: `API call error: ${String(err)}`,
+        tokenUsage: {
+          total: 0,
+          prompt: 0,
+          completion: 0,
+        },
+      };
+    }
+    logger.debug(`\tAzure API response (embeddings): ${JSON.stringify(data)}`);
+
+    try {
+      const embedding = data?.data?.[0]?.embedding;
+      if (!embedding) {
+        throw new Error('No embedding returned');
+      }
+      const ret = {
+        embedding,
+        tokenUsage: cached
+          ? { cached: data.usage.total_tokens, total: data.usage.total_tokens }
+          : {
+              total: data.usage.total_tokens,
+              prompt: data.usage.prompt_tokens,
+              completion: data.usage.completion_tokens,
+            },
+      };
+      return ret;
+    } catch (err) {
+      return {
+        error: `API response error: ${String(err)}: ${JSON.stringify(data)}`,
+        tokenUsage: cached
+          ? {
+              cached: data.usage.total_tokens,
+              total: data.usage.total_tokens,
+            }
+          : {
+              total: data?.usage?.total_tokens,
+              prompt: data?.usage?.prompt_tokens,
+              completion: data?.usage?.completion_tokens,
+            },
+      };
+    }
+  }
+}
